Add to list from another page
Using the constructor, I pass data from the class addPerson(new class)
to the class ContainerList. The class ContainerList stores the data
to be added, and after clicking the button adds a new Person.
You can add the same data as many times as you like. When you re-enter
the add page, the data that you added last time is erased.
It is necessary to change the approach and not to pass data from class
to class through constructors, but to implement the add method Person in
the class AddPerson directly to the list