Work with provider
To implement the addition of elements from another page, I decided to
study the provider package at a basic level. Several hours of torment
behind, but the goal is not achieved. But there are no errors

This class takes a list of instances of the class Person using
the constructor and converts it to a widget ListView

class ListPersonWrapper extends StatelessWidget {
  List<Person> pList = [];
  ListPersonWrapper({Key? key, required this.pList}) : super(key: key);
  @override
  Widget build(BuildContext context) {
    return ListView.separated(
        itemCount: person.personList.length,
        separatorBuilder: (BuildContext context, int index)
        => const Divider(),
        itemBuilder: (BuildContext context, int index) {
          return ListTile(
            title: Text(pList[index].name),
            subtitle: Text(pList[index].companies),
          );
        });
  }
}
Here is the class constructor ListPersonWrapper with the help of the
getter acquires the value of the list, which is tracked for changes
withthe help of notifyListeners(). There is a final transformation
into a widget and display on the screen.

class _HomePageState extends State<HomePage> {
  @override
  Widget build(BuildContext context) {
    Widget PersonL =
        ListPersonWrapper(pList: context.watch<PersonList>().getData);

    return Scaffold(
      appBar: AppBar(
        title: Text("Avqvq"),
      ),
      body: PersonL,
    );
  }